% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{A real data set on Dozer Cycle Times}
\format{
A data frame with 72 observations
}
\usage{
data1
}
\description{
The data set from {AbouRizk, S.M., Halpin, D.W., Wilson, J. R. (1994). \emph{Fitting beta distributions based on sample data.} Journal of Construction Engineering and Management 120(2), 288–305.}
consist of 72 observations for Dozer Cycle Times
}
\keyword{data}
\keyword{set}
