% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{data2}
\alias{data2}
\title{A real data set on the height of five-year-old British boys}
\format{
A data frame with 99 observations
}
\usage{
data2
}
\description{
The data set presents the height of 99 five-year-old British boys in cm downloaded from {http://www.mas.ncl.ac.uk/njnsm/medfac/docs/intro.pdf.}
}
\keyword{data}
\keyword{set}
