% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLF.stat.R
\name{MLF.stat}
\alias{MLF.stat}
\title{Modified Lilliefors Goodness-of-Fit Normality Test}
\usage{
MLF.stat(x)
}
\arguments{
\item{x}{a numeric vector of data values, the number of which must be greater than 4.}
}
\value{
The function returns the value of the modified Lilliefors goodness-of-fit normality test statistic.
}
\description{
Calculates the value of the modified Lilliefors goodness-of-fit normality test statistic.
}
\details{
The modified Lilliefors goodness-of-fit normality test statistic, see formula (5) in the article.
}
\examples{
MLF.stat(rnorm(33, mean = 0, sd = 2))
MLF.stat(data1)

}
\references{
{Sulewski, P. (2019). \emph{Modified Lilliefors Goodness-of-fit Test for Normality.} Communications in Statistics - Simulation and Computation 51(3), 1199-1219.}
}
\author{
Piotr Sulewski, \email{piotr.sulewski@apsl.edu.pl}, Pomeranian University in Slupsk.
}
