% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSGoft.R
\name{PSGoft}
\alias{PSGoft}
\alias{PSGoft-package}
\title{The list of package functions and their demonstration}
\description{
The \strong{PSGoft} package puts into practice the modified Lilliefors goodness-of-fit normality test. This modification consists in varying a formula of calculating the empirical distribution function.
Values of constants a, b in the formula depend on values of sample skewness and excess kurtosis,
which is recommended in order to increase the power of the LF test.
}
\section{Data sets in the package}{


\code{\link{data1}}

\code{\link{data2}}
}

\section{Functions for the modified Lilliefors goodness-of-fit normality test}{


\code{\link{MLF.stat}}

\code{\link{MLF.pvalue}}

\code{\link{MLF.test}}
}

